strukt/requirements.txt
# Core framework dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional dependencies for examples and helpers
# Uncomment the ones you need:

# LLM deps
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0

# For vector stores (if using memory)
upstash-vector>=0.1.0

# For async support (already included in Python 3.7+)
asyncio

# For development/testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
